packages/rose-web/pyproject.toml
[project]
name = "rose-web"
version = "0.1.3"
description = "ROSE Web Application"
readme = "README.md"
requires-python = ">=3.12,<4.0"
authors = [{ name = "Garden Variety AI" }]
license = { text = "MIT" }
keywords = ["llm", "ai", "model-conversion", "utilities"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "openai>=1.88.0",
    "openai-agents>=0.0.19",
    "sse-starlette>=2.1.3",
    "greenlet>=3.2.2",
    "jinja2>=3.1.6",
    "openai-chatkit>=1.0.2",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.24",
    "aiosqlite>=0.21.0",
]

[project.scripts]
rose-web = "rose_web.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/rose_web"]
